+++ packages/router/src/adapters/contract/contract.ts
@@ -13,7 +13,7 @@ import {
   InvariantTransactionData,
   TransactionData,
 } from "@connext/nxtp-utils";
-import { BigNumber, constants, Contract, providers, utils } from "ethers/lib/ethers";
+import { BigNumber, constants, Contract, providers, utils } from "ethers";
 import { Evt } from "evt";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
 import { TransactionManager as TTransactionManager } from "@connext/nxtp-contracts/typechain";
@@ -273,6 +273,20 @@ export const prepareRouterContract = async (
     signature,
   ]);
 
+  const onchainTx = {
+    to: routerContractAddress,
+    data: encodedData,
+    value: constants.Zero,
+    chainId,
+    from: wallet.address,
+  };
+  if (useRelayer) {
+    // If we are will be using relayers below, check to make sure the transaction is valid first (before relaying it)
+    // by running an estimateGas check. This method will throw a TransactionReverted error (with the contract error code)
+    // if the transaction would fail on chain.
+    await txService.getGasEstimate(chainId, onchainTx);
+  }
+
   // 1. Prepare tx using relayer if chain is supported by gelato.
   if (useRelayer && isChainSupportedByGelato(chainId)) {
     logger.info("Router contract prepare: sending using Gelato relayer", requestContext, methodContext, {
@@ -366,13 +380,7 @@ export const prepareRouterContract = async (
   logger.info("Router contract prepare: sending using txservice", requestContext, methodContext, { prepareParams });
 
   const receipt = await txService.sendTx(
-    {
-      to: routerContractAddress,
-      data: encodedData,
-      value: constants.Zero,
-      chainId,
-      from: wallet.address,
-    },
+    onchainTx,
     requestContext,
   );
 
@@ -475,6 +483,22 @@ export const fulfillRouterContract = async (
     signature,
   ]);
 
+  const onchainTx = {
+    to: routerContractAddress,
+    data: encodedData,
+    value: constants.Zero,
+    chainId,
+    from: wallet.address,
+  };
+  if (useRelayer) {
+    // If we are will be using relayers below, check to make sure the transaction is valid first (before relaying it)
+    // by running an estimateGas check. This method will throw a TransactionReverted error (with the contract error code)
+    // if the transaction would fail on chain.
+    // TODO: Would be nice to recycle the gasLimit we get back from this call in the event that we end up
+    // using txservice.
+    await txService.getGasEstimate(chainId, onchainTx);
+  }
+
   // 1. Prepare tx using relayer if chain is supported by gelato.
   if (useRelayer && isChainSupportedByGelato(chainId)) {
     logger.info("Router contract fulfill: sending using Gelato relayer", requestContext, methodContext, {
@@ -568,13 +592,7 @@ export const fulfillRouterContract = async (
   // 3. If all of the above failed or was otherwise not supported, use txservice to send the transaction.
   logger.info("Router contract fulfill: sending using txservice", requestContext, methodContext, { fulfillParams });
   const receipt = await txService.sendTx(
-    {
-      to: routerContractAddress,
-      data: encodedData,
-      value: constants.Zero,
-      chainId,
-      from: wallet.address,
-    },
+    onchainTx,
     requestContext,
   );
 
@@ -657,6 +675,20 @@ export const cancelRouterContract = async (
     routerRelayerFee,
     signature,
   ]);
+
+  const onchainTx = {
+    to: routerContractAddress,
+    data: encodedData,
+    value: constants.Zero,
+    chainId,
+    from: wallet.address,
+  };
+  if (useRelayer) {
+    // If we are will be using relayers below, check to make sure the transaction is valid first (before relaying it)
+    // by running an estimateGas check. This method will throw a TransactionReverted error (with the contract error code)
+    // if the transaction would fail on chain.
+    await txService.getGasEstimate(chainId, onchainTx);
+  }
 
   // 1. Prepare tx using relayer if chain is supported by gelato.
   if (useRelayer && isChainSupportedByGelato(chainId)) {
@@ -753,13 +785,7 @@ export const cancelRouterContract = async (
 
   logger.info("Router contract cancel: sending using txservice", requestContext, methodContext, { cancelParams });
   const receipt = await txService.sendTx(
-    {
-      to: routerContractAddress,
-      data: encodedData,
-      value: constants.Zero,
-      chainId,
-      from: wallet.address,
-    },
+    onchainTx,
     requestContext,
   );
 
@@ -847,6 +873,20 @@ export const removeLiquidityRouterContract = async (
     signature,
   ]);
 
+  const onchainTx = {
+    to: routerContractAddress,
+    data: encodedData,
+    value: constants.Zero,
+    chainId,
+    from: wallet.address,
+  };
+  if (useRelayer) {
+    // If we are will be using relayers below, check to make sure the transaction is valid first (before relaying it)
+    // by running an estimateGas check. This method will throw a TransactionReverted error (with the contract error code)
+    // if the transaction would fail on chain.
+    await txService.getGasEstimate(chainId, onchainTx);
+  }
+
   // 1. Prepare tx using relayer if chain is supported by gelato.
   if (useRelayer && isChainSupportedByGelato(chainId)) {
     logger.info("Router contract removeLiquidity: sending using Gelato relayer", requestContext, methodContext, {
@@ -923,16 +963,7 @@ export const removeLiquidityRouterContract = async (
     amount,
     assetId,
   });
-  return await txService.sendTx(
-    {
-      to: routerContractAddress,
-      data: encodedData,
-      value: constants.Zero,
-      chainId,
-      from: wallet.address,
-    },
-    requestContext,
-  );
+  return await txService.sendTx(onchainTx, requestContext);
 };
 
 export const addLiquidityForTransactionManager = async (