+++ packages/sdk/src/sdkBase.ts
@@ -987,11 +987,11 @@ export class NxtpSdkBase {
       throw new ChainNotConfigured(txData.receivingChainId, Object.keys(this.config.chainConfig));
     }
 
-    const fulfillTxProm = this.waitFor(SubgraphEvents.ReceiverTransactionFulfilled, FULFILL_TIMEOUT, (data) => {
-      return data.txData.transactionId === params.txData.transactionId;
-    });
-
     if (useRelayers) {
+      const fulfillTxProm = this.waitFor(SubgraphEvents.ReceiverTransactionFulfilled, FULFILL_TIMEOUT, (data) => {
+        return data.txData.transactionId === params.txData.transactionId;
+      });
+
       if (isChainSupportedByGelato(txData.receivingChainId)) {
         this.logger.info("Fulfilling using Gelato Relayer", requestContext, methodContext);
         const deployedContract = this.config.chainConfig[txData.receivingChainId].transactionManagerAddress!;
@@ -1156,7 +1156,7 @@ export class NxtpSdkBase {
   }
 
   public async querySubgraph(chainId: number, query: string): Promise<any> {
-    this.subgraph.query(chainId, query);
+    return this.subgraph.query(chainId, query);
   }
 
   /**